Copied files from the old repository, purposefully losing history.
Source repository:
 https://bitbucket.org/adam4813/nls-engine
Source Revision information:
 Revision: 312
 Changeset: 651d856e67e6bcc99f8e0a507e24cffe73b1f7ca [651d856e67e6]
 Author: "Ricky Curtice" <kf6kjg+bitbucket@gmail.com>
 Date: May 6, 2012 9:26:14 PM PDT

The reason for this loss is twofold: we have almost completely changed the engine; we made a huge mistake back in Saturn V - one we knew would bite us - that of incorporating our libraries into our source tree.  While the former reason isn't sufficient to lose history in such a huge stroke as this, the latter is.  The repository had a huge excess of files that had nothing to do with core functionality: Boost's headers took more space and file count than our entire engine on its own.  It had originally been included so that we didn't have to have the headache of instructing people on how to install Boost first.  However, with the conversion to CMake, I've now invalidated that reason: Boost, along with any other library that the engine needs, is downloaded and set up automatically by CMake.  This allowed me to remove all of those pesky files - from everywhere but the mercurial history of course.  The result is this major loss of 312 revisions.

That said, it's not a total loss: we don't plan on deleting the old repository any time soon.  In all likelihood it will just sit until BitBucket pulls the rug - by then it should have served its purpose.