Changed Boost from "versioned" to just "tagged"
This not only removes the version tags, which I couldn't care one way or the other about, but it also has the feature that it removes the compiler version - a big thing for cross-platform ease of use.
I had attempted to use wildcards to catch the compiler name, however SET() doesn't do globbing, and the result was the * being passed to the linker as a literal "*" character.
Attempting to use the completely untagged mode also caused a problem - both the debug and the optimized editions wound up with the same file name - not good.  So "tagged" mode wins.