Fixed a bug where a libary wouldn't be extracted in certain cases.
When libraries were upgraded, or the downloaded file was deleted, AND CMake had previously decompressed an archive for that library, the decompression process wouldn't happen.  Workarounds are to either uncheck the LIB_x_ALREADY_EXTRACTED flag or clean the CMake cache so that CMake would catch it next time you reconfigured.

The solution was to simply force the LIB_x_ALREADY_EXTRACTED flag to false when the file was downloaded so the system would then automatically decompress the archive and copy the files.