Upgraded Boost and added compilation flags for using clang/LLVM with the libc++ standard library.
Warning! Please read the following posts before attempting to compile on OSX:
* http://www.gamedev.net/topic/628077-osx-clang-and-getcwd/
* http://rwcproductions.com/node/26

As inferred by the above warning I'm working on getting the codebase to compile on Mac OSX.  I'm testing under Xcode 4.3/4.4 on both 10.7 (Lion) and 10.8 (Mountain Lion).  However it's still not as easy as it should be: the project settings need some manual tweaking in Xcode before it will compile.  For instance, even though the -std=c++11 and -stdlib=libc++ flags are set, Xcode still needs those manually selected in the GUI.  I'll be hunting a fix for that shortly.

Found out that I needed to force C++11 mode in GLM under clang, as the library is checking compilers, not features.  Silly library.

Pretty sure AS_MAX_PORTABILITY isn't needed any more… If it is, it should be placed in the cases where it IS needed.  If not, then next time someone goes near this code please remove that line.