Swapped out offset of for asOFFSET to clear up a warning about using offset of on a non-POD type.
However the fundamental issue isn't resolved.

New issue
Summary: Properties of non-POD types are being registered with offsetof/asOFFST
asOFFSET is nothing more than a hack to clear out the warning GCC/Clang emits when offset of is used against a non-POD type.  It would be better to explore the idea of registering accessors: IIRC the script interface doesn't change, just the registration.

Here's what the AS manual has to say:
"It is also possible to expose properties through property accessors, which are a pair of class methods with prefixes get_ and set_ for getting and setting the property value. These methods should be registered with RegisterObjectMethod. This is especially useful when the offset of the property cannot be determined, or if the type of the property is not registered in the script and some translation must occur, i.e. from char* to string."
The documentation for the property accessor registration is here: http://www.angelcode.com/angelscript/sdk/docs/manual/doc_script_class_prop.html